pyproject.toml
[project]
name = "cpctf25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "ptrlib",
    "pwntools>=4.14.1",
    "pycryptodome>=3.22.0",
]

[tool.uv.sources]
ptrlib = { git = "https://github.com/ptr-yudai/ptrlib" }
